Modified firebase stuff
